refactor TemporaryDirectory to avoid import from conda.compat
